+++ src/muncher/monster/type.js
@@ -22,22 +22,26 @@ export function getType(monster, DDB_CONFIG) {
     "custom": ""
   };
 
-  const typeName = DDB_CONFIG.monsterTypes.find((c) => monster.typeId == c.id).name.toLowerCase();
-  if (CONFIG.DND5E.creatureTypes[typeName]) result.value = typeName;
-
-  result.subtype = DDB_CONFIG.monsterSubTypes
-    .filter((c) => monster.subTypes.includes(c.id))
-    .map((c) => c.name)
-    .join(", ");
-
   if (monster.swarm) {
     // result.swarm.isSwarm = true;
     // result.swarm.size = getSizeFromId(monster.swarm.sizeId, DDB_CONFIG).value;
     result.swarm = getSizeFromId(monster.swarm.sizeId, DDB_CONFIG).value;
   }
 
+  const type = DDB_CONFIG.monsterTypes.find((c) => monster.typeId == c.id);
+  if (!type) {
+    result.custom = "Unknown";
+    return result;
+  }
 
-  // To Do : custom type
+  const typeName = type.name.toLowerCase();
+
+  if (CONFIG.DND5E.creatureTypes[typeName]) result.value = typeName;
+
+  result.subtype = DDB_CONFIG.monsterSubTypes
+    .filter((c) => monster.subTypes.includes(c.id))
+    .map((c) => c.name)
+    .join(", ");
 
   return result;