+++ src/parser/spells/damage.js
@@ -19,14 +19,20 @@ export function getDamage(data, spell) {
   };
 
   const globalDamageHints = game.settings.get("ddb-importer", "use-damage-hints");
+  const damageRestrictionHints = game.settings.get("ddb-importer", "add-damage-restrictions-to-hints");
+  const hintOrRestriction = globalDamageHints || damageRestrictionHints;
+
   // damage
   const attacks = data.definition.modifiers.filter((mod) => mod.type === "damage");
   if (attacks.length !== 0) {
     const cantripBoost = data.definition.level === 0 && !!data.flags.ddbimporter.dndbeyond.cantripBoost;
     attacks.forEach((attack) => {
-      const damageHint = globalDamageHints ? `[${attack.subType}]` : "";
+      const restriction = damageRestrictionHints && attack.restriction && attack.restriction !== "" ? attack.restriction : "";
+      const hintAndRestriction = globalDamageHints && restriction !== "" ? " - " : "";
+      const damageHint = globalDamageHints ? attack.subType : "";
+      const damageTag = hintOrRestriction ? `[${damageHint}${hintAndRestriction}${restriction}]` : "";
       const addMod = attack.usePrimaryStat || cantripBoost ? " + @mod" : "";
-      let diceString = utils.parseDiceString(attack.die.diceString, addMod, damageHint).diceString;
+      let diceString = utils.parseDiceString(attack.die.diceString, addMod, damageTag).diceString;
       result.parts.push([diceString, attack.subType]);
     });
 
@@ -41,8 +47,11 @@ export function getDamage(data, spell) {
   if (heals.length !== 0) {
     const healingBonus = (spell.flags.ddbimporter.dndbeyond.healingBoost) ? ` + ${spell.flags.ddbimporter.dndbeyond.healingBoost} + @item.level` : "";
     heals.forEach((heal) => {
-      const damageHint = globalDamageHints ? `[healing]` : "";
-      const healValue = (heal.die.diceString) ? `${heal.die.diceString}${damageHint}` : heal.die.fixedValue;
+      const restriction = damageRestrictionHints && heal.restriction && heal.restriction !== "" ? heal.restriction : "";
+      const hintAndRestriction = globalDamageHints && restriction !== "" ? " - " : "";
+      const damageHint = globalDamageHints ? "healing" : "";
+      const damageTag = hintOrRestriction ? `[${damageHint}${hintAndRestriction}${restriction}]` : "";
+      const healValue = (heal.die.diceString) ? `${heal.die.diceString}${damageTag}` : heal.die.fixedValue;
       const diceString = heal.usePrimaryStat
         ? `${healValue} + @mod${healingBonus}`
         : `${healValue}${healingBonus}`;