+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Image upload directory is now set and created to a sensible default.
 * Adventure Importer Journals with tables now offer a Roll on Table button.
 * Heat Metal spell parsed as a spell attack.
+* Option to add damage restrictions to damage hint tag.
 
 # 1.0.9